Second Part of OBB and physics
Currently the table can't be placed anywhere else then 0,0,0 since the
OBBs aren't rotated, scaled or transposed

There is also a small bug that occurs if an object hits the edge of the
angled, scaled cube. The sphere will speed up noticeably. That's because
the the speed vector for the second / third collision test is already
transformed and so there is an additional transformation that pushes the
"gravity" speed towards the movement direction.